refactor(vc): extract vc_validation.go from vc_service.go
Move VC verification methods and result types to a dedicated file.
Pure file move: no symbol renames or logic changes.

Refs #415
